docs: update FormFiled documentation
