+++ doc-site/pages/components/form-field.mdx
@@ -9,7 +9,7 @@ import Imports from '../../components/imports';
 import playgroundConfig from '../../examples/form-field/playground-config';
 import BasicUsage from '../../examples/form-field/basic-usage';
 import States from '../../examples/form-field/states';
-import Required from '../../examples/form-field/required';
+import RequiredIndicator from '../../examples/form-field/required-indicator';
 import Inverted from '../../examples/form-field/inverted';
 
 export default Layout;
@@ -30,8 +30,8 @@ Use this utility component in conjunction with [Input](/components/input), [Text
   <States />
 </Example>
 
-<Example title="Required" path="form-field/required">
-  <Required />
+<Example title="Required Indicator" path="form-field/required-indicator">
+  <RequiredIndicator />
 </Example>
 
 <Example title="Inverted" path="form-field/inverted" darkPreview>
